+++ lib/docusign_monitor/configuration.rb
@@ -1,7 +1,7 @@
 =begin
 #Monitor API
 
-#Use the DocuSign Monitor API to receive a data feed containing atypical security events within your DocuSign account. This data goes directly to an integrated application or website.
+#An API for an integrator to access the features of DocuSign Monitor
 
 OpenAPI spec version: v2.0
 Contact: devcenter@docusign.com