+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # DocuSign Monitor Ruby Client Changelog
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
+## [v1.1.0] - Monitor API v2.0-1.1.0 - 2022-04-11
+### Changed
+- Added support for version v2.0-1.1.0 of the DocuSign Monitor API.
+- Updated the SDK release version.
+
 ## [v1.0.0] - DocuSign Monitor API v1 - 06/24/2021
 ### Changed
-- First GA version of Monitor API, supports DocuSign Monitor API
+- First version of Monitor API, supports DocuSign Monitor API